+++ serverside_challenge_1/challenges/yukikawamura/electricity-rate-simulator/electricity_rate_simulator/core.py
@@ -1,10 +1,7 @@
 import json
-import logging
-from pathlib import Path
 
 from pydantic import ValidationError
 
-from .model import PlanContract, PlanUsage, ProFile, UserData
 from .exception import (
     ElectricSimulateClientError,
     ElectricSimulateProviderError,
@@ -12,23 +9,8 @@ from .exception import (
     NotFoundContractError,
     NotFoundProviderError,
 )
-
-BASE_DIR = Path(__file__).parents[0]
-DATA_DIR = BASE_DIR.joinpath("data")
-PROVIDER_DIR = DATA_DIR.joinpath("provider")
-
-
-def setup_logging(debug_mode=False):
-    lgr = logging.getLogger("uvicorn.app")
-    log_format = "%(asctime)s:[%(levelname)s] %(message)s"
-    stream_handler = logging.StreamHandler()
-    stream_handler.setFormatter(log_format)
-    level = logging.DEBUG if debug_mode else logging.INFO
-
-    lgr.setLevel(level)
-    lgr.addHandler(stream_handler)
-    return lgr
-
+from .model import PlanContract, PlanUsage, ProFile, UserData
+from .utils import PROVIDER_DIR, setup_logging
 
 lgr = setup_logging()
 
@@ -83,7 +65,7 @@ class ElectricSimulator(object):
             )
         except ValidationError as e:
             raise ElectricSimulationError(e)
-            
+
         except ElectricSimulateProviderError as e:
             raise e
 
@@ -108,7 +90,7 @@ class ElectricSimulator(object):
             }
 
             return simulation
-        except (ElectricSimulateProviderError, ElectricSimulateClientError) as e:
+        except ElectricSimulateClientError as e:
             lgr.exception(e)
             raise e
