+++ serverside_challenge_1/challenges/yukikawamura/electricity-rate-simulator/tests/test_api.py
@@ -14,7 +14,8 @@ def test_electric_simulations_api():
     params = {"contract": 10, "usage": 100}
     res = test_client.get("/simulations", params=params)
     assert res.status_code == 200
-    assert res.json() == [{
+    assert res.json() == [
+        {
             "provider": "東京電力エナジーパートナー",
             "plan": "従量電灯B",
             "price": "2274円",
@@ -23,7 +24,8 @@ def test_electric_simulations_api():
             "provider": "Loopでんき",
             "plan": "おうちプラン",
             "price": "2640円",
-        }]
+        },
+    ]
 
 
 def test_electric_simulations_api_invalid_contract():
@@ -32,8 +34,8 @@ def test_electric_simulations_api_invalid_contract():
     params = {"contract": contract, "usage": usage}
     res = test_client.get("/simulations", params=params)
     assert res.status_code == 400
-    assert res.json() == {'detail': f'Invalid number of contract: {contract}'}
-    
+    assert res.json() == {"detail": f"Invalid number of contract: {contract}"}
+
 
 def test_electric_simulations_api_invalid_usage():
     contract = 10
@@ -43,3 +45,26 @@ def test_electric_simulations_api_invalid_usage():
     assert res.status_code == 400
     assert res.json() == {"detail": f"Invalid number of usage: {usage}"}
 
+
+def test_electric_simulations_api_invalid_contract_value():
+    contract = "test"
+    usage = 100
+    params = {"contract": contract, "usage": usage}
+    res = test_client.get("/simulations", params=params)
+    assert res.status_code == 422
+    assert (
+        res.text
+        == '{"detail":[{"type":"int_parsing","loc":["query","contract"],"msg":"Input should be a valid integer, unable to parse string as an integer","input":"test","url":"https://errors.pydantic.dev/2.6/v/int_parsing"}]}'
+    )
+
+
+def test_electric_simulations_api_invalid_usage_value():
+    contract = 10
+    usage = "test"
+    params = {"contract": contract, "usage": usage}
+    res = test_client.get("/simulations", params=params)
+    assert res.status_code == 422
+    assert (
+        res.text
+        == '{"detail":[{"type":"int_parsing","loc":["query","usage"],"msg":"Input should be a valid integer, unable to parse string as an integer","input":"test","url":"https://errors.pydantic.dev/2.6/v/int_parsing"}]}'
+    )
